+++ plugin-dev/Source/Sentry/Private/Android/SentryEventAndroid.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Sentry. All Rights Reserved.
 
 #include "SentryEventAndroid.h"
-
+#include "SentryIdAndroid.h"
 #include "SentryMessageAndroid.h"
 
 #include "Infrastructure/SentryConvertersAndroid.h"
@@ -21,11 +21,18 @@ SentryEventAndroid::SentryEventAndroid(jobject event)
 
 void SentryEventAndroid::SetupClassMethods()
 {
+	GetIdMethod = GetMethod("getEventId", "()Lio/sentry/protocol/SentryId;");
 	SetMessageMethod = GetMethod("setMessage", "(Lio/sentry/protocol/Message;)V");
 	GetMessageMethod = GetMethod("getMessage", "()Lio/sentry/protocol/Message;");
 	SetLevelMethod = GetMethod("setLevel", "(Lio/sentry/SentryLevel;)V");
 	GetLevelMethod = GetMethod("getLevel", "()Lio/sentry/SentryLevel;");
 	IsCrashMethod = GetMethod("isCrashed", "()Z");
+}
+
+TSharedPtr<ISentryId> SentryEventAndroid::GetId() const
+{
+	auto id = CallObjectMethod<jobject>(GetIdMethod);
+	return MakeShareable(new SentryIdAndroid(*id));
 }
 
 void SentryEventAndroid::SetMessage(const FString& message)